+++ activerecord/lib/active_record/tasks/postgresql_database_tasks.rb
@@ -30,7 +30,7 @@ module ActiveRecord
       end
 
       def collation
-        connection.collate
+        connection.collation
       end
 
       def purge